Rename saturn_api to graph_api
Better reflects that this file provides the C API for the Graph object
specifically, not a general Saturn API.
