+++ rust/saturn-sys/src/lib.rs
@@ -1,2 +1,2 @@
 pub mod conversions;
-pub mod saturn_api;
+pub mod graph_api;
